+++ scripts/release/utils/index.ts
@@ -3,5 +3,4 @@ export * from "./package-info";
 export * from "./shell";
 export * from "./git";
 export * from "./github";
-export * from "./config";
 export * from "./changelog-parser";